-Fixing scence

+++ a3/starter/utils.c
@@ -399,21 +399,25 @@ double under_root=coe_b*coe_b-(double)4*coe_a*coe_c;
  free(normal);
  free(world_normal);
 }
-void convert_xyz_to_cube_uv(float x, float y, float z, int *index, float *u, float *v)
+void convert_xyz_to_cube_uv(double x, double y, double z, int *index, double *u, double *v)
 {
 /*
 A cube texture indexes six texture maps from 0 to 5 in
 order Positive X, Negative X, Positive Y, Negative Y, Positive Z, Negative Z
 */
-  float absX = fabs(x);
-  float absY = fabs(y);
-  float absZ = fabs(z);
+    //std::cout<<" x: "<<x<<" y: "<<y<<" z "<<z<<std::endl;
+
+  double absX = fabs(x);
+  double absY = fabs(y);
+  double absZ = fabs(z);
 
   int isXPositive = x > 0 ? 1 : 0;
   int isYPositive = y > 0 ? 1 : 0;
   int isZPositive = z > 0 ? 1 : 0;
 
   float maxAxis, uc, vc;
+
+  //std::cout<<" absx: "<<absX<<" absY: "<<absY<<" absZ "<<absZ<<std::endl;
 
   // POSITIVE X
   if (isXPositive && absX >= absY && absX >= absZ) {
@@ -469,6 +473,8 @@ order Positive X, Negative X, Positive Y, Negative Y, Positive Z, Negative Z
     vc = y;
     *index = 5;
   }
+
+   //std::cout<<" uc: "<<uc<<" vc "<<vc<<std::endl;
 
   // Convert range from -1 to 1 to 0 to 1
   *u = 0.5f * (uc / maxAxis + 1.0f);
